requirements.txt
python-dotenv
hikari
hikari-tanjun
hikari-yuyo
lavaplayer